third_party/blink/web_tests/external/wpt/css/css-sizing/contain-intrinsic-size/auto-012-expected.txt

This is a testharness.js-based test.
PASS Sizing normally
PASS Sizing with c-i-s fallback
PASS Still sizing with c-i-s fallback
FAIL Sizing with last remembered size assert_equals: clientWidth expected 100 but got 50
FAIL Still sizing with last remembered size assert_equals: clientWidth expected 100 but got 50
PASS Sizing with new c-i-s fallback
FAIL Sizing with new last remembered size assert_equals: clientWidth expected 150 but got 50
Harness: the test ran to completion.
